refactor(routes): remove Route suffixes from file names
