+++ server/src/routes/affixes.ts
