+++ server/src/routes/chatCommands.ts
