+++ server/src/routes/configs.ts
