+++ server/src/routes/files.ts
