+++ server/src/routes/initRoutes.ts
@@ -1,23 +1,23 @@
 import { Express } from "express";
-import auth from "./authRoute";
-import chatCommandsRouter from "./chatCommandsRoute";
-import configsRouter from "./configsRoute";
-import messagesRouter from "./messagesRoute";
-import redemptionsRouter from "./redemptionsRoute";
-import triggersRouter from "./triggersRoute";
-import timersRouter from "./timersRoute";
-import usersRouter from "./usersRoute";
-import streamSessionsRouter from "./streamSessionsRoute";
-import messageCategoriesRouter from "./messageCategoriesRoute";
-import overlayRouter from "./overlayRoute";
-import widgetsRouter from "./widgetsRoute";
-import moodsRouter from "./moodsRoute";
-import tagsRouter from "./tagsRoute";
-import affixesRouter from "./affixesRoute";
-import filesRouter from "./filesRoute";
-import songsRouter from "./songsRoute";
-import achievementsRouter from "./achievementsRoute";
-import youtubeRouter from "./youtubeRoute";
+import auth from "./auth";
+import chatCommandsRouter from "./chatCommands";
+import configsRouter from "./configs";
+import messagesRouter from "./messages";
+import redemptionsRouter from "./redemptions";
+import triggersRouter from "./triggers";
+import timersRouter from "./timers";
+import usersRouter from "./users";
+import streamSessionsRouter from "./streamSessions";
+import messageCategoriesRouter from "./messageCategories";
+import overlayRouter from "./overlays";
+import widgetsRouter from "./widgets";
+import moodsRouter from "./moods";
+import tagsRouter from "./tags";
+import affixesRouter from "./affixes";
+import filesRouter from "./files";
+import songsRouter from "./songs";
+import achievementsRouter from "./achievements";
+import youtubeRouter from "./youtube";
 
 export const initRoutes = (app: Express) => {
   app.use("/achievements", achievementsRouter);
