+++ server/src/routes/triggers.ts
